Move demo view-model and theme to :demo-core module
