+++ demo-core/build.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Backbase R&D B.V.
+ * Copyright 2021 Backbase R&D B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 apply plugin: 'org.jetbrains.kotlin.android'
 
 android {
@@ -22,13 +22,11 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
 
     defaultConfig {
-        applicationId 'com.backbase.deferredresources.demo'
         minSdkVersion rootProject.ext.minSdk
         targetSdkVersion rootProject.ext.targetSdk
-        versionCode 1
-        versionName rootProject.ext.libraryVersion
+        versionName version
 
-        vectorDrawables.useSupportLibrary = true
+        consumerProguardFiles 'consumer-rules.pro'
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -36,12 +34,7 @@ android {
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
-    }
-
-    buildFeatures {
-        viewBinding true
     }
 
     compileOptions {
@@ -51,18 +44,13 @@ android {
 
     kotlinOptions {
         jvmTarget = JavaVersion.VERSION_1_8
+        freeCompilerArgs += ['-Xexplicit-api=strict']
     }
 }
 
 dependencies {
-    implementation(project(':demo-core'))
+    api(project(':deferred-resources'))
+    api("com.google.android.material:material:1.3.0")
 
-    implementation(project(':deferred-resources'))
-    implementation(project(':deferred-resources-view-extensions'))
-
-    implementation("androidx.activity:activity-ktx:1.2.3")
-    implementation("androidx.viewpager2:viewpager2:1.0.0")
-
-    implementation(libs.androidx.appCompat)
     implementation(libs.androidx.coreKtx)
 }
