+++ demo-core/src/main/res/drawable/ic_fire_24.xml
