+++ demo-core/src/main/res/drawable/ic_flower_24.xml
