+++ demo-core/src/main/res/values-ar/plurals.xml
