+++ demo-core/src/main/res/values-night-v27/colors.xml
