+++ demo-core/src/main/res/values-night/colors.xml
