+++ demo-core/src/main/res/values-v27/colors.xml
