+++ demo-core/src/main/res/values/plurals.xml
