+++ demo/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
       android:label="@string/app_name"
       android:roundIcon="@mipmap/ic_launcher_round"
       android:supportsRtl="true"
-      android:theme="@style/AppTheme">
+      android:theme="@style/Theme.DeferredResourcesDemo">
 
     <activity android:name=".DemoActivity">
         <intent-filter>
