+++ demo/src/main/java/com/backbase/deferredresources/demo/DemoActivity.kt
@@ -17,15 +17,14 @@
 package com.backbase.deferredresources.demo
 
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.backbase.deferredresources.demo.databinding.DemoBinding
-import com.backbase.deferredresources.demo.model.SamplesViewModel
+import com.backbase.deferredresources.demo.core.SamplesViewModel
 import com.google.android.material.tabs.TabLayoutMediator
 
 class DemoActivity : AppCompatActivity() {
 
-    private val viewModel: SamplesViewModel by viewModels()
+    private val viewModel: SamplesViewModel by samplesViewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
